requirements.txt
pyserial
cryptocode
opencv-python
qrcode
pytesseract
fpdf
pillow
flask
flask-restful
flask-cors
mysql.connector